Remove CommunicateMesh
